ai-service-python/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
opencv-python
numpy
Pillow